tlshd: use gnutls_psk_allocate_{client,server}_credentials2
There is a merge request to gnutls (#11939) to implement
gnutls_psk_allocate_client_credentials2() and
gnutls_psk_allocate_server_credentials2(). These new functions
allow to specify the binder algorithm, to overcome the defect
in gnutls to always use SHA256 when calculating the binder value.
So update the code to use these functions if available.

Signed-off-by: Hannes Reinecke <hare@suse.de>